Final Commit

+++ AlertToCareAPI/Repo/IcuConfigrationRepository.cs
@@ -1,6 +1,7 @@
 ﻿using AlertToCare.Data;
 using AlertToCareAPI.Database;
 using AlertToCareAPI.Models;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
@@ -40,6 +41,16 @@ namespace AlertToCareAPI.Repo
             {
                 throw new ArgumentNullException(nameof(icu));
             }
+            //If beds of the Icu are occupied throw an exception
+            var occupiedBeds = _context.BedsInfo.FromSqlRaw($"SELECT * FROM BedsInfo WHERE IcuId = {icu.Id} AND IsOccupid = 1").ToList();
+
+            if (occupiedBeds.Count() > 0)
+            {
+                throw new Exception("ICU cann't be removed still got some occupied beds !!");
+            }
+            //Remove the beds of the corresponding ICU from the bed table
+            _context.BedsInfo.FromSqlRaw($"DELETE FROM BedsInfo WHERE IcuId = {icu.Id}");
+
             _context.IcusInfo.Remove(icu);
         }
